RESTAPI/requirements.txt
Flask-CORS
python-dotenv
psycopg2-binary